タブをReactのActivityベースにする
