+++ src/pages/DetailPage/DetailPage.tsx
@@ -1,0 +1,25 @@
+import { Tabs } from "./components/Tabs";
+import { Activity } from "react";
+import { PaymentBreakdownView } from "./components/PaymentBreakdownView/PaymentBreakdownView";
+import { PaymentView } from "./components/PaymentView/PaymentView";
+
+type Props = {
+  fileName: string;
+  activeTab: "payments" | "breakdown";
+};
+
+export function DetailPage({ fileName, activeTab }: Props) {
+  return (
+    <div className="flex flex-col gap-6">
+      <Tabs fileName={fileName} activeTab={activeTab} />
+
+      <Activity mode={activeTab === "payments" ? "visible" : "hidden"}>
+        <PaymentView fileName={fileName} />
+      </Activity>
+
+      <Activity mode={activeTab === "breakdown" ? "visible" : "hidden"}>
+        <PaymentBreakdownView fileName={fileName} />
+      </Activity>
+    </div>
+  );
+}
