+++ src/pages/DetailPage/components/PaymentBreakdownView/PayviewBarChart.tsx
