--- src/routes/breakdown.$fileName.tsx
@@ -1,12 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { PaymentsBreakdownPage } from "../pages/PaymentsBreakdownPage";
-
-export const Route = createFileRoute("/breakdown/$fileName")({
-  component: _PaymentsBreakdownPage,
-});
-
-function _PaymentsBreakdownPage() {
-  // eslint-disable-next-line react-hooks/rules-of-hooks
-  const { fileName } = Route.useParams();
-  return <PaymentsBreakdownPage fileName={fileName} />;
-}
