+++ src/routes/payments.$fileName.tsx
@@ -1,12 +1,22 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { PaymentPage } from "../pages/PaymentPage";
+import { DetailPage } from "../pages/DetailPage/DetailPage";
+import { z } from "zod";
+
+const SearchSchema = z.object({
+  tab: z.enum(["payments", "breakdown"]).default("payments"),
+});
 
 export const Route = createFileRoute("/payments/$fileName")({
   component: _PaymentPage,
+  validateSearch: SearchSchema,
 });
 
 function _PaymentPage() {
   // eslint-disable-next-line react-hooks/rules-of-hooks
   const { fileName } = Route.useParams();
-  return <PaymentPage fileName={fileName} />;
+
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  const { tab } = Route.useSearch();
+
+  return <DetailPage fileName={fileName} activeTab={tab} />;
 }
